requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.1
fastapi-utils==0.2.1
PyMuPDF==1.23.8
python-docx==1.1.0
openai==1.93.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
typing-inspect==0.9.0